Detector Calibration Simulation/requirements.txt
# Detector Calibration Simulation Requirements
# Core scientific computing and data analysis packages

# Numerical computing
numpy>=1.21.0
scipy>=1.7.0

# Data analysis and manipulation
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Image processing (optional, for advanced features)
Pillow>=8.0.0
scikit-image>=0.18.0

# Wavelet transforms (optional, for wavelet denoising)
PyWavelets>=1.1.0

# HDF5 support (optional, for large datasets)
h5py>=3.1.0

# Configuration file parsing
PyYAML>=6.0

# Progress bars (optional, for long simulations)
tqdm>=4.62.0

# Interactive plotting (optional)
plotly>=5.0.0

# Statistical analysis (optional)
statsmodels>=0.12.0

# Machine learning (optional, for advanced calibration)
scikit-learn>=1.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Testing framework (optional, for development)
pytest>=6.0.0
pytest-cov>=2.12.0

# Code formatting and linting (optional, for development)
black>=21.0.0
flake8>=3.9.0

# Type checking (optional, for development)
mypy>=0.910
